fix lint errors in configgen (#13758)

+++ package/batocera/core/batocera-configgen/configgen/configgen/generators/bigpemu/bigpemuGenerator.py
@@ -257,7 +257,7 @@ class BigPEmuGenerator(Generator):
         config["BigPEmuConfig"]["Video"]["ScreenScaling"] = 5
         config["BigPEmuConfig"]["Video"]["DisplayWidth"] = gameResolution["width"]
         config["BigPEmuConfig"]["Video"]["DisplayHeight"] = gameResolution["height"]
-        config["BigPEmuConfig"]["Video"]["DisplayFrequency"] = int(round(float(videoMode.getRefreshRate())))
+        config["BigPEmuConfig"]["Video"]["DisplayFrequency"] = round(float(videoMode.getRefreshRate()))
 
         # User selections
         config["BigPEmuConfig"]["Video"]["VSync"] = system.config.get("bigpemu_vsync", 1)
